move memoery mapped file to pixels-common.
