+++ pixels-cache/src/test/java/io/pixelsdb/pixels/cache/TestMemFileWrite.java
@@ -19,6 +19,7 @@
  */
 package io.pixelsdb.pixels.cache;
 
+import io.pixelsdb.pixels.common.physical.natives.MemoryMappedFile;
 import org.junit.Test;
 
 import java.io.File;
