+++ pixels-common/src/test/java/io/pixelsdb/pixels/common/physical/natives/TestDirect.java
@@ -17,7 +17,7 @@
  * License along with Pixels.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package io.pixelsdb.pixels.common.physical.direct;
+package io.pixelsdb.pixels.common.physical.natives;
 
 import org.junit.Test;
